Add install subcommand
This command is a wrapper script for virsh, virt-install and
virt-clone which is intended to limit usecases for simplicity. Currently
there are three usecases and the first `install` is only implemented in
this commit. The rest will be implemented in the next commits.

* install: alias for virt-install with default variables.
* clone: create a clone of VM prepared by install subcommand.
* remove: shoutdown and remove a set of specified VMs.

The usage is like this one. Notice that each of them is implemented as a
subcommand.

  $ sudo src/pyvirt install
  or
  $ sudo src/pyvirt clone VM_NAME